Added option to show/hide additional holes on sides of the mount.
Improved key slots for mounting so that centeral one can be one way up or the other and added a bottom set that allows the mount to be mounted the other way around (power switch / usb port on left).

Modified centeral cable cutouts to accomodate mini-trunking rather than just USB cable, this is useful if mini-trunking is fitted going up so the mount can still be removed (or fitted) from the wall with the key mounts.